sd に応じた色を返す処理をhooksに切り出し
