+++ app/_hooks/index.ts
@@ -1,1 +1,2 @@
+export * from "./useAccuracyColor";
 export * from "./useBpmCalculator";
